requirements.txt
PySimpleGUI~=4.18.1
scikit-image~=0.22.0
numpy~=1.25.2
opencv-python~=4.8.1.78
Pillow~=10.0.0